fix converter bug (effect-related)
Only remember the effect parameter if there actually is an effect.

Reapply the effect when the instrument changes, even if the effect
+ param is the same as before, because the NES playback engine resets
the effect when a new instrument is set.

This makes Sweet Child O' Mine convert fine again, and keeps
Legend of the Black Shawarma medley & Smooth Criminal still working,
at least.